+++ CHANGELOG.md
@@ -75,3 +75,4 @@
 * Replace `PL` newtype with `Tagged` newtype from `tagged` package
 * Replace `Message'` with `Way` and include it to `Capacity` with `Send` constructor
 * Remove `user_id` field from `Contact` datatype
+* Define `Persistable` instances for `Location` datatype